Keep the next daemon and config hotspots from re-inflating after the runtime refactor wave
This follow-up continues the hotspot reduction work by extracting the ACP CLI
surface out of daemon/lib, splitting onboarding selection and screen-spec
helpers out of onboard_cli, and moving channel default/validation helpers out
of config/channels.

The goal is not new behavior; it is to keep the recently cleaned runtime path
reviewable while preserving the same onboarding, ACP, and channel-config
contracts under full verification.

Constraint: The slice had to stay behavior-preserving because PR #1218 is already under review
Constraint: Shared cargo target caches in this repo are currently unreliable across worktrees, so verification had to use isolated CARGO_TARGET_DIR state
Rejected: Leave daemon/lib, onboard_cli, and channel_config just under or over the hotspot thresholds | the next maturity wave would immediately accumulate more pressure in the same files
Rejected: Take a broad onboarding or channel-config redesign in one pass | too much semantic risk for a review-followup refactor
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep ACP CLI/read-model rendering, onboarding presentation helpers, and channel default/validation helpers in focused submodules instead of regrowing the top-level hotspot files
Tested: CARGO_TARGET_DIR=/tmp/loongclaw-acp-refactor-nlcgLa cargo clippy --workspace --all-targets --all-features -- -D warnings; CARGO_TARGET_DIR=/tmp/loongclaw-acp-refactor-nlcgLa cargo test --workspace --all-features -- --test-threads=1; ./scripts/check_architecture_boundaries.sh; ./scripts/check_dep_graph.sh; ./scripts/generate_architecture_drift_report.sh docs/releases/architecture-drift-2026-04.md; ./scripts/check_architecture_drift_freshness.sh docs/releases/architecture-drift-2026-04.md; git diff --check
Not-tested: remote CI
Related: #1217
Related: #1218
