+++ crates/daemon/src/onboard_cli.rs
@@ -8,7 +8,6 @@ use std::time::Duration;
 
 use dialoguer::console::{Term, user_attended};
 use dialoguer::theme::ColorfulTheme;
-use dialoguer::{Confirm, Error as DialoguerError, FuzzySelect, Input, Select};
 use kernel::ToolCoreRequest;
 use loongclaw_app as mvp;
 use loongclaw_contracts::SecretRef;
@@ -62,6 +61,14 @@ use std::fs;
 #[cfg(test)]
 use time::OffsetDateTime;
 
+#[path = "onboard_select.rs"]
+mod select_support;
+
+use self::select_support::*;
+#[path = "onboard_screen_specs.rs"]
+mod screen_spec_support;
+
+use self::screen_spec_support::*;
 pub use crate::onboard_finalize::{
     OnboardingAction, OnboardingActionKind, OnboardingDomainOutcome, OnboardingSuccessSummary,
     backup_existing_config, build_onboarding_success_summary,
@@ -567,362 +574,6 @@ fn rich_prompt_theme() -> ColorfulTheme {
 
 fn rich_prompt_term() -> Term {
     Term::stdout()
-}
-
-fn render_select_option_item(option: &SelectOption) -> String {
-    let mut rendered = option.label.clone();
-    if !option.description.trim().is_empty() {
-        rendered.push_str(" - ");
-        rendered.push_str(option.description.trim());
-    }
-    if option.recommended {
-        rendered.push_str(" (recommended)");
-    }
-    rendered
-}
-
-fn map_rich_prompt_error(action: &str, error: DialoguerError) -> String {
-    let error: io::Error = error.into();
-    if error.kind() == io::ErrorKind::Interrupted {
-        return "onboarding cancelled: prompt aborted".to_owned();
-    }
-    format!("{action} failed: {error}")
-}
-
-fn prompt_with_default_rich(label: &str, default: &str) -> CliResult<String> {
-    let term = rich_prompt_term();
-    prompt_with_default_rich_on(&term, label, default)
-}
-
-fn prompt_with_default_rich_on(term: &Term, label: &str, default: &str) -> CliResult<String> {
-    let theme = rich_prompt_theme();
-    let value = Input::<String>::with_theme(&theme)
-        .with_prompt(label)
-        .default(default.to_owned())
-        .report(false)
-        .interact_text_on(term)
-        .map_err(|error| map_rich_prompt_error("interactive prompt", error))?;
-    let value = ensure_onboard_input_not_cancelled(value)?;
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Ok(default.to_owned());
-    }
-    Ok(trimmed.to_owned())
-}
-
-fn prompt_required_rich(label: &str) -> CliResult<String> {
-    let term = rich_prompt_term();
-    prompt_required_rich_on(&term, label)
-}
-
-fn prompt_required_rich_on(term: &Term, label: &str) -> CliResult<String> {
-    let theme = rich_prompt_theme();
-    let value = Input::<String>::with_theme(&theme)
-        .with_prompt(label)
-        .report(false)
-        .interact_text_on(term)
-        .map_err(|error| map_rich_prompt_error("interactive prompt", error))?;
-    let value = ensure_onboard_input_not_cancelled(value)?;
-    Ok(value.trim().to_owned())
-}
-
-fn prompt_allow_empty_rich(label: &str) -> CliResult<String> {
-    let term = rich_prompt_term();
-    prompt_allow_empty_rich_on(&term, label)
-}
-
-fn prompt_allow_empty_rich_on(term: &Term, label: &str) -> CliResult<String> {
-    let theme = rich_prompt_theme();
-    let value = Input::<String>::with_theme(&theme)
-        .with_prompt(label)
-        .allow_empty(true)
-        .report(false)
-        .interact_text_on(term)
-        .map_err(|error| map_rich_prompt_error("interactive prompt", error))?;
-    let value = ensure_onboard_input_not_cancelled(value)?;
-    Ok(value.trim().to_owned())
-}
-
-fn prompt_confirm_rich(message: &str, default: bool) -> CliResult<bool> {
-    let term = rich_prompt_term();
-    let theme = rich_prompt_theme();
-    Confirm::with_theme(&theme)
-        .with_prompt(message)
-        .default(default)
-        .report(false)
-        .interact_on_opt(&term)
-        .map_err(|error| map_rich_prompt_error("interactive confirmation", error))?
-        .ok_or_else(|| "onboarding cancelled: prompt aborted".to_owned())
-}
-
-fn select_one_rich(
-    label: &str,
-    options: &[SelectOption],
-    default: Option<usize>,
-    interaction_mode: SelectInteractionMode,
-) -> CliResult<usize> {
-    let default = validate_select_one_state(options.len(), default)?;
-    let items = options
-        .iter()
-        .map(render_select_option_item)
-        .collect::<Vec<_>>();
-    let term = rich_prompt_term();
-    let theme = rich_prompt_theme();
-    let selection = match interaction_mode {
-        SelectInteractionMode::List => {
-            let prompt = Select::with_theme(&theme)
-                .with_prompt(label)
-                .items(&items)
-                .report(false);
-            let prompt = if let Some(idx) = default {
-                prompt.default(idx)
-            } else {
-                prompt
-            };
-            prompt
-                .interact_on_opt(&term)
-                .map_err(|error| map_rich_prompt_error("interactive selection", error))?
-        }
-        SelectInteractionMode::Search => {
-            let prompt = FuzzySelect::with_theme(&theme)
-                .with_prompt(label)
-                .items(&items)
-                .report(false);
-            let prompt = if let Some(idx) = default {
-                prompt.default(idx)
-            } else {
-                prompt
-            };
-            prompt
-                .interact_on_opt(&term)
-                .map_err(|error| map_rich_prompt_error("interactive model search", error))?
-        }
-    };
-    selection.ok_or_else(|| "onboarding cancelled: prompt aborted".to_owned())
-}
-
-fn summarize_select_option_description(detail_lines: &[String]) -> String {
-    detail_lines
-        .iter()
-        .map(String::as_str)
-        .map(str::trim)
-        .filter(|line| !line.is_empty())
-        .collect::<Vec<_>>()
-        .join(" | ")
-}
-
-fn select_options_from_screen_options(options: &[OnboardScreenOption]) -> Vec<SelectOption> {
-    options
-        .iter()
-        .map(|option| SelectOption {
-            label: option.label.clone(),
-            slug: option.key.clone(),
-            description: summarize_select_option_description(&option.detail_lines),
-            recommended: option.recommended,
-        })
-        .collect()
-}
-
-fn tui_choices_from_screen_options(options: &[OnboardScreenOption]) -> Vec<TuiChoiceSpec> {
-    options
-        .iter()
-        .map(|option| TuiChoiceSpec {
-            key: option.key.clone(),
-            label: option.label.clone(),
-            detail_lines: option.detail_lines.clone(),
-            recommended: option.recommended,
-        })
-        .collect()
-}
-
-fn select_screen_option(
-    ui: &mut impl OnboardUi,
-    label: &str,
-    options: &[OnboardScreenOption],
-    default_key: Option<&str>,
-) -> CliResult<usize> {
-    let select_options = select_options_from_screen_options(options);
-    let default_idx =
-        default_key.and_then(|key| options.iter().position(|option| option.key == key));
-    ui.select_one(
-        label,
-        &select_options,
-        default_idx,
-        SelectInteractionMode::List,
-    )
-}
-
-fn build_onboard_entry_screen_options(options: &[OnboardEntryOption]) -> Vec<OnboardScreenOption> {
-    options
-        .iter()
-        .enumerate()
-        .map(|(index, option)| OnboardScreenOption {
-            key: (index + 1).to_string(),
-            label: option.label.to_owned(),
-            detail_lines: vec![option.detail.clone()],
-            recommended: option.recommended,
-        })
-        .collect()
-}
-
-fn build_starting_point_selection_screen_options(
-    sorted_candidates: &[ImportCandidate],
-    width: usize,
-) -> Vec<OnboardScreenOption> {
-    let mut options = sorted_candidates
-        .iter()
-        .enumerate()
-        .map(|(index, candidate)| OnboardScreenOption {
-            key: (index + 1).to_string(),
-            label: onboard_starting_point_label(Some(candidate.source_kind), &candidate.source),
-            detail_lines: summarize_starting_point_detail_lines(candidate, width),
-            recommended: matches!(
-                candidate.source_kind,
-                crate::migration::ImportSourceKind::RecommendedPlan
-            ),
-        })
-        .collect::<Vec<_>>();
-    options.push(OnboardScreenOption {
-        key: "0".to_owned(),
-        label: crate::onboard_presentation::start_fresh_option_label().to_owned(),
-        detail_lines: start_fresh_starting_point_detail_lines(),
-        recommended: false,
-    });
-    options
-}
-
-fn build_onboard_shortcut_screen_options(
-    shortcut_kind: OnboardShortcutKind,
-) -> Vec<OnboardScreenOption> {
-    vec![
-        OnboardScreenOption {
-            key: "1".to_owned(),
-            label: shortcut_kind.primary_label().to_owned(),
-            detail_lines: vec![crate::onboard_presentation::shortcut_continue_detail().to_owned()],
-            recommended: true,
-        },
-        OnboardScreenOption {
-            key: "2".to_owned(),
-            label: crate::onboard_presentation::adjust_settings_label().to_owned(),
-            detail_lines: vec![crate::onboard_presentation::shortcut_adjust_detail().to_owned()],
-            recommended: false,
-        },
-    ]
-}
-
-fn build_existing_config_write_screen_options() -> Vec<OnboardScreenOption> {
-    vec![
-        OnboardScreenOption {
-            key: "o".to_owned(),
-            label: "Replace existing config".to_owned(),
-            detail_lines: vec!["overwrite the current file with this onboarding draft".to_owned()],
-            recommended: false,
-        },
-        OnboardScreenOption {
-            key: "b".to_owned(),
-            label: "Create backup and replace".to_owned(),
-            detail_lines: vec![
-                "save a timestamped .bak copy first, then write the new config".to_owned(),
-            ],
-            recommended: true,
-        },
-        OnboardScreenOption {
-            key: "c".to_owned(),
-            label: "Cancel".to_owned(),
-            detail_lines: vec!["leave the existing config untouched".to_owned()],
-            recommended: false,
-        },
-    ]
-}
-
-fn validate_select_one_state(
-    options_len: usize,
-    default: Option<usize>,
-) -> CliResult<Option<usize>> {
-    if options_len == 0 {
-        return Err("no selection options available".to_owned());
-    }
-    if let Some(idx) = default
-        && idx >= options_len
-    {
-        return Err(format!(
-            "default selection index {idx} out of range 0..{}",
-            options_len - 1
-        ));
-    }
-    Ok(default)
-}
-
-fn select_option_input_slug(option: &SelectOption) -> &str {
-    if option.slug == ONBOARD_CUSTOM_MODEL_OPTION_SLUG {
-        "custom"
-    } else {
-        option.slug.as_str()
-    }
-}
-
-fn parse_select_one_input(trimmed: &str, options: &[SelectOption]) -> Option<usize> {
-    if let Ok(selected) = trimmed.parse::<usize>()
-        && (1..=options.len()).contains(&selected)
-    {
-        return Some(selected - 1);
-    }
-
-    let direct_match = options.iter().position(|option| {
-        option.slug.eq_ignore_ascii_case(trimmed)
-            || select_option_input_slug(option).eq_ignore_ascii_case(trimmed)
-    });
-
-    if direct_match.is_some() {
-        return direct_match;
-    }
-
-    parse_prompt_personality_select_input(trimmed, options)
-}
-
-fn render_select_one_invalid_input_message(options: &[SelectOption]) -> String {
-    format!(
-        "invalid selection. enter a number between 1 and {}, or one of: {}",
-        options.len(),
-        options
-            .iter()
-            .map(select_option_input_slug)
-            .collect::<Vec<_>>()
-            .join(", ")
-    )
-}
-
-fn resolve_select_one_eof(default: Option<usize>) -> CliResult<usize> {
-    default.ok_or_else(|| {
-        "onboarding cancelled: stdin closed while waiting for required selection".to_owned()
-    })
-}
-
-fn parse_prompt_personality_select_input(trimmed: &str, options: &[SelectOption]) -> Option<usize> {
-    let prompt_personality_options = select_options_are_prompt_personalities(options);
-
-    if !prompt_personality_options {
-        return None;
-    }
-
-    let personality = parse_prompt_personality(trimmed)?;
-    let canonical_slug = prompt_personality_id(personality);
-
-    options
-        .iter()
-        .position(|option| option.slug.eq_ignore_ascii_case(canonical_slug))
-}
-
-fn select_options_are_prompt_personalities(options: &[SelectOption]) -> bool {
-    for option in options {
-        let parsed_personality = parse_prompt_personality(&option.slug);
-
-        if parsed_personality.is_none() {
-            return false;
-        }
-    }
-
-    true
 }
 
 fn print_lines(ui: &mut impl OnboardUi, lines: impl IntoIterator<Item = String>) -> CliResult<()> {
@@ -4410,130 +4061,14 @@ fn render_shortcut_default_choice_footer_line(shortcut_kind: OnboardShortcutKind
     render_default_choice_footer_line("1", shortcut_kind.default_choice_description())
 }
 
+fn tui_header_style(style: OnboardHeaderStyle) -> TuiHeaderStyle {
+    match style {
+        OnboardHeaderStyle::Compact => TuiHeaderStyle::Compact,
+    }
+}
+
 pub fn render_onboarding_risk_screen_lines(width: usize) -> Vec<String> {
     render_onboarding_risk_screen_lines_with_style(width, false)
-}
-
-fn render_onboarding_risk_screen_lines_with_style(
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let copy = crate::onboard_presentation::risk_screen_copy();
-    let footer_lines = append_escape_cancel_hint(vec![render_default_choice_footer_line(
-        "n",
-        copy.default_choice_description,
-    )]);
-    let spec = TuiScreenSpec {
-        header_style: TuiHeaderStyle::Brand,
-        subtitle: Some(copy.subtitle.to_owned()),
-        title: Some(copy.title.to_owned()),
-        progress_line: None,
-        intro_lines: vec!["review the trust boundary before writing any config".to_owned()],
-        sections: vec![
-            TuiSectionSpec::Callout {
-                tone: TuiCalloutTone::Warning,
-                title: Some("what onboarding can do".to_owned()),
-                lines: vec![
-                    "LoongClaw can invoke tools and read local files when enabled.".to_owned(),
-                    "Keep credentials in environment variables, not in prompts.".to_owned(),
-                    "Prefer allowlist-style tool policy for shared environments.".to_owned(),
-                ],
-            },
-            TuiSectionSpec::Narrative {
-                title: Some("recommended baseline".to_owned()),
-                lines: vec![
-                    "start with the narrowest tool scope that still lets you verify first success"
-                        .to_owned(),
-                    "you can widen channels, models, and local automation after doctor and review"
-                        .to_owned(),
-                ],
-            },
-        ],
-        choices: vec![
-            TuiChoiceSpec {
-                key: "y".to_owned(),
-                label: copy.continue_label.to_owned(),
-                detail_lines: vec![copy.continue_detail.to_owned()],
-                recommended: false,
-            },
-            TuiChoiceSpec {
-                key: "n".to_owned(),
-                label: copy.cancel_label.to_owned(),
-                detail_lines: vec![copy.cancel_detail.to_owned()],
-                recommended: false,
-            },
-        ],
-        footer_lines,
-    };
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn build_onboard_shortcut_screen_spec(
-    shortcut_kind: OnboardShortcutKind,
-    config: &mvp::config::LoongClawConfig,
-    import_source: Option<&str>,
-    include_choices: bool,
-) -> TuiScreenSpec {
-    let mut snapshot_lines = Vec::new();
-    if let Some(source) = import_source {
-        let starting_point_label = onboard_starting_point_label(None, source);
-        snapshot_lines.push(onboard_display_line(
-            "- starting point: ",
-            &starting_point_label,
-        ));
-    }
-    snapshot_lines.extend(build_onboard_review_digest_display_lines(config));
-    let snapshot_title = if import_source.is_some() {
-        "detected starting point snapshot"
-    } else {
-        "current setup snapshot"
-    };
-
-    let choices = if include_choices {
-        tui_choices_from_screen_options(&build_onboard_shortcut_screen_options(shortcut_kind))
-    } else {
-        Vec::new()
-    };
-    let default_choice_footer_line = render_shortcut_default_choice_footer_line(shortcut_kind);
-    let footer_lines = append_escape_cancel_hint(vec![default_choice_footer_line]);
-
-    TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: Some(shortcut_kind.subtitle().to_owned()),
-        title: Some(shortcut_kind.title().to_owned()),
-        progress_line: None,
-        intro_lines: Vec::new(),
-        sections: vec![
-            TuiSectionSpec::Narrative {
-                title: Some(snapshot_title.to_owned()),
-                lines: snapshot_lines,
-            },
-            TuiSectionSpec::Callout {
-                tone: TuiCalloutTone::Success,
-                title: Some("fast lane".to_owned()),
-                lines: vec![shortcut_kind.summary_line().to_owned()],
-            },
-        ],
-        choices,
-        footer_lines,
-    }
-}
-
-fn render_preflight_summary_screen_lines_with_style(
-    checks: &[OnboardCheck],
-    width: usize,
-    flow_style: ReviewFlowStyle,
-    color_enabled: bool,
-) -> Vec<String> {
-    let progress_line = flow_style.progress_line();
-
-    render_preflight_summary_screen_lines_with_progress(
-        checks,
-        width,
-        progress_line.as_str(),
-        color_enabled,
-    )
 }
 
 pub fn render_write_confirmation_screen_lines(
@@ -4576,126 +4111,6 @@ pub fn render_detected_setup_write_confirmation_screen_lines(
         ReviewFlowStyle::QuickDetectedSetup,
         false,
     )
-}
-
-fn render_write_confirmation_screen_lines_with_style(
-    config_path: &str,
-    warnings_kept: bool,
-    width: usize,
-    flow_style: ReviewFlowStyle,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec = build_write_confirmation_screen_spec(config_path, warnings_kept, flow_style);
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn build_onboard_choice_screen_spec(
-    header_style: OnboardHeaderStyle,
-    subtitle: &str,
-    title: &str,
-    step: Option<(GuidedOnboardStep, GuidedPromptPath)>,
-    intro_lines: Vec<String>,
-    options: Vec<OnboardScreenOption>,
-    footer_lines: Vec<String>,
-    show_escape_cancel_hint: bool,
-) -> TuiScreenSpec {
-    let resolved_subtitle = screen_subtitle(subtitle);
-    let resolved_progress_line =
-        step.map(|(step, guided_prompt_path)| step.progress_line(guided_prompt_path));
-    let resolved_footer_lines = if show_escape_cancel_hint {
-        append_escape_cancel_hint(footer_lines)
-    } else {
-        footer_lines
-    };
-    let resolved_choices = tui_choices_from_screen_options(&options);
-
-    TuiScreenSpec {
-        header_style: tui_header_style(header_style),
-        subtitle: resolved_subtitle,
-        title: Some(title.to_owned()),
-        progress_line: resolved_progress_line,
-        intro_lines,
-        sections: Vec::new(),
-        choices: resolved_choices,
-        footer_lines: resolved_footer_lines,
-    }
-}
-
-fn build_onboard_input_screen_spec(
-    title: &str,
-    step: GuidedOnboardStep,
-    guided_prompt_path: GuidedPromptPath,
-    context_lines: Vec<String>,
-    hint_lines: Vec<String>,
-) -> TuiScreenSpec {
-    let resolved_footer_lines = append_escape_cancel_hint(hint_lines);
-    let progress_line = step.progress_line(guided_prompt_path);
-
-    TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: None,
-        title: Some(title.to_owned()),
-        progress_line: Some(progress_line),
-        intro_lines: context_lines,
-        sections: Vec::new(),
-        choices: Vec::new(),
-        footer_lines: resolved_footer_lines,
-    }
-}
-
-fn build_write_confirmation_screen_spec(
-    config_path: &str,
-    warnings_kept: bool,
-    flow_style: ReviewFlowStyle,
-) -> TuiScreenSpec {
-    let mut intro_lines = Vec::new();
-    let config_line = format!("- config: {config_path}");
-    let status_line =
-        crate::onboard_presentation::write_confirmation_status_line(warnings_kept).to_owned();
-
-    intro_lines.push(config_line);
-    intro_lines.push(status_line);
-
-    let choices = vec![
-        TuiChoiceSpec {
-            key: "y".to_owned(),
-            label: crate::onboard_presentation::write_confirmation_label().to_owned(),
-            detail_lines: vec![crate::onboard_presentation::write_confirmation_detail().to_owned()],
-            recommended: false,
-        },
-        TuiChoiceSpec {
-            key: "n".to_owned(),
-            label: crate::onboard_presentation::write_confirmation_cancel_label().to_owned(),
-            detail_lines: vec![
-                crate::onboard_presentation::write_confirmation_cancel_detail().to_owned(),
-            ],
-            recommended: false,
-        },
-    ];
-
-    let default_choice_line = render_default_choice_footer_line(
-        "y",
-        crate::onboard_presentation::write_confirmation_default_choice_description(),
-    );
-    let footer_lines = append_escape_cancel_hint(vec![default_choice_line]);
-
-    TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: None,
-        title: Some(crate::onboard_presentation::write_confirmation_title().to_owned()),
-        progress_line: Some(flow_style.progress_line()),
-        intro_lines,
-        sections: Vec::new(),
-        choices,
-        footer_lines,
-    }
-}
-
-fn tui_header_style(style: OnboardHeaderStyle) -> TuiHeaderStyle {
-    match style {
-        OnboardHeaderStyle::Compact => TuiHeaderStyle::Compact,
-    }
 }
 
 fn screen_subtitle(subtitle: &str) -> Option<String> {
